split structure creator module
